Clean up Pause Menu + Loading Screen (#75)
* remove loading screen video

* remove muse branding from pause menu

* move pausemenu to separate folder, add customization options

* add shadow to loading screen
